// front-final/src/app/models/photographer.model.ts
export interface Photographer {
  firstName: string;
  lastName: string;
  city: string;
  country: string;
  email: string;
  specialityField1: string;
  specialityField2: string;
  specialityField3: string;
  specialityField4: string;
  specialityField5: string;
  specialityField6: string;
  specialityField7: string;
  specialityField8: string;
  gender: string;
  password: string;
  contactNo: string;
  imgUrl1: string;
  imgUrl2: string;
  imgUrl3: string;
  imgUrl4: string;
  imgUrl5: string;
}